% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{cfm}
\alias{cfm}
\title{Centre for Macroeconomics expert panel survey results}
\format{Data frame.}
\source{
The CFM Surveys \url{http://cfmsurvey.org/}
}
\description{
Likert votes by 58 economists on several years' worth of policy surveys 
conducted by the Centre for Macroeconomics
}
