% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{eigm}
\alias{eigm}
\title{Initiative on Global Markets European expert panel survey results}
\format{Data frame.}
\source{
IGM Forum
  \url{http://www.igmchicago.org/european-economic-experts-panel}
}
\description{
European counterpart to \code{\link{igm}}
}

