% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{igm}
\alias{igm}
\title{Initiative on Global Markets expert panel survey results}
\format{Data frame.}
\source{
IGM Forum \url{http://www.igmchicago.org/igm-economic-experts-panel}
}
\description{
Likert votes by 50 economists on several years' worth of policy surveys 
conducted by the Initiative on Global Markets at the University of Chicago's 
Booth School of Business
}

