% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{planetmoney}
\alias{planetmoney}
\title{Planet Money presidential primary proposals assessment}
\format{Data frame.}
\source{
Jess Jiang, "Economists On Candidates' Proposals: Mostly Bad". Planet
  Money, 26 February 2016 
  \url{http://www.npr.org/sections/money/2016/02/26/468298576/economists-on-candidates-proposals-mostly-bad}
}
\description{
Frequency dataset of likert votes by 22 economists of several candidates' 
policy proposals in a survey conducted by NPR's Planet Money team during the 
2015-6 presidential primary season
}
